+++ tests/urls.py
@@ -1,6 +1,9 @@
+from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
+    # admin
+    path("admin/", admin.site.urls),
     # indieweb
-    path("indieweb/", include("indieweb.urls"))
+    path("indieweb/", include("indieweb.urls")),
 ]